requirements.txt
torch
tqdm
numpy
pandas
transformers
datasets
huggingface-hub
python-dotenv
faiss-cpu
scikit-learn
seaborn